change datatype

+++ Colour.cs
@@ -4,12 +4,12 @@ namespace ClassesTask1
 {
     class Colour
     {
-        public float r;
-        public float g;
-        public float b;
+        public int r;
+        public int g;
+        public int b;
         public string name;
 
-        public Colour(float r, float g, float b, string name)
+        public Colour(int r, int g, int b, string name)
         {
             this.r = r;
             this.g = g;
